requirements.txt
Django == 4.2.5
djangorestframework==3.14.0
drf-spectacular == 0.26.4
django-cors-headers == 4.2.0
Pillow == 10.0.0